[ALPS03892298] m4u:fixed auto merge conflict
[Detail]
fixed auto merge conflict

MTK-Commit-Id: 8abe163e5332ec5181f9f734ab9fc8886df44ce4

CR-Id: ALPS03892298
Feature:[Module]Kernel Maintenance

+++ drivers/misc/mediatek/m4u/2.0/m4u_mva.c
@@ -43,7 +43,7 @@ void m4u_mvaGraph_init(void *priv_reserve)
 	spin_lock_irqsave(&gMvaGraph_lock, irq_flags);
 	memset(mvaGraph, 0, sizeof(short) * (MVA_MAX_BLOCK_NR + 1));
 	memset(mvaInfoGraph, 0, sizeof(void *) * (MVA_MAX_BLOCK_NR + 1));
-	mvaGraph[0] = 1 | MVA_BUSY_MASK;
+	mvaGraph[0] = 1 | (short)MVA_BUSY_MASK;
 	mvaInfoGraph[0] = priv_reserve;
 	mvaGraph[1] = MVA_MAX_BLOCK_NR;
 	mvaInfoGraph[1] = priv_reserve;
